+++ src/main/java/com/github/dawnwords/jacoco/badge/JacocoResultParser.java
@@ -30,7 +30,7 @@ class JacocoResultParser {
       throws Exception {
     final Path reportPath = Paths.get(setting.getJacocoReportPath());
     final HashMap<String, JacocoBadgePercentageResult> results = new HashMap<>();
-    final Map<String, Integer> limit = Optional.ofNullable(setting.getLimit()).orElse(
+    final Map<String, Double> limit = Optional.ofNullable(setting.getLimit()).orElse(
         Collections.emptyMap()).entrySet().stream().collect(Collectors.toMap(
         e -> e.getKey().toUpperCase(), Entry::getValue
     ));
